modularized the main file
